Use reusable Status componetn for ProposedNotReacted
