+++ front/app/containers/InitiativesShow/ReactionControl/Status/components/DisabledReasonTooltip.tsx
@@ -1,0 +1,82 @@
+import { Icon, colors, fontSizes } from '@citizenlab/cl2-component-library';
+import { darken } from 'polished';
+import React from 'react';
+import styled from 'styled-components';
+import { InitiativePermissionsDisabledReason } from 'hooks/useInitiativesPermissions';
+import messages from '../../messages';
+import globalMessages from 'utils/messages';
+import { MessageDescriptor } from 'react-intl';
+import { FormattedMessage } from 'utils/cl-intl';
+import { InitiativeDisabledReason } from 'api/initiative_action_descriptors/types';
+
+const TooltipContent = styled.div<{ inMap?: boolean }>`
+  display: flex;
+  align-items: center;
+  padding: ${(props) => (props.inMap ? '0px' : '15px')};
+`;
+
+const TooltipContentIcon = styled(Icon)`
+  flex: 0 0 24px;
+  margin-right: 1rem;
+`;
+
+const TooltipContentText = styled.div`
+  flex: 1 1 auto;
+  color: ${colors.textPrimary};
+  font-size: ${fontSizes.base}px;
+  line-height: normal;
+  font-weight: 400;
+  overflow-wrap: break-word;
+  word-wrap: break-word;
+  word-break: break-word;
+
+  a,
+  button {
+    color: ${colors.teal};
+    font-size: ${fontSizes.base}px;
+    line-height: normal;
+    font-weight: 400;
+    text-align: left;
+    text-decoration: underline;
+    white-space: normal;
+    overflow-wrap: break-word;
+    word-wrap: break-word;
+    word-break: break-all;
+    word-break: break-word;
+    hyphens: auto;
+    display: inline;
+    padding: 0px;
+    margin: 0px;
+    cursor: pointer;
+    transition: all 100ms ease-out;
+
+    &:hover {
+      color: ${darken(0.15, colors.teal)};
+      text-decoration: underline;
+    }
+  }
+`;
+
+interface Props {
+  disabledReason: InitiativeDisabledReason;
+}
+
+const disabledMessages: {
+  [key in InitiativePermissionsDisabledReason]: MessageDescriptor;
+} = {
+  not_permitted: messages.votingNotPermitted,
+  not_in_group: globalMessages.notInGroup,
+};
+
+const DisabledReasonTooltip = ({ disabledReason }: Props) => {
+  return (
+    <TooltipContent id="tooltip-content" className="e2e-disabled-tooltip">
+      <TooltipContentIcon name="lock" ariaHidden />
+      <TooltipContentText>
+        <FormattedMessage {...disabledMessages[disabledReason]} />
+      </TooltipContentText>
+    </TooltipContent>
+  );
+};
+
+export default DisabledReasonTooltip;
