+++ front/app/containers/InitiativesShow/ReactionControl/Status/components/VoteButtons.tsx
@@ -3,11 +3,14 @@ import Button from 'components/UI/Button';
 import messages from '../../messages';
 import { FormattedMessage } from 'utils/cl-intl';
 import { StatusComponentProps } from '..';
+import Tippy from '@tippyjs/react';
+import DisabledReasonTooltip from './DisabledReasonTooltip';
 
 interface Props {
   userReacted: StatusComponentProps['userReacted'];
   onCancelReaction: StatusComponentProps['onCancelReaction'];
   onReaction: StatusComponentProps['onReaction'];
+  disabledReason: StatusComponentProps['disabledReason'];
   voteButtonId?: string;
   cancelVoteButtonId?: string;
   cancelReactionDisabled: boolean;
@@ -17,10 +20,15 @@ const VoteButtons = ({
   userReacted,
   onCancelReaction,
   onReaction,
+  disabledReason,
   voteButtonId,
   cancelVoteButtonId,
   cancelReactionDisabled,
 }: Props) => {
+  const tippyContent = disabledReason ? (
+    <DisabledReasonTooltip disabledReason={disabledReason} />
+  ) : null;
+
   return (
     <>
       {userReacted ? (
@@ -35,15 +43,23 @@ const VoteButtons = ({
           <FormattedMessage {...messages.voted} />
         </Button>
       ) : (
-        <Button
-          buttonStyle="primary"
-          iconSize="20px"
-          icon="vote-ballot"
-          onClick={onReaction}
-          id={voteButtonId}
+        <Tippy
+          disabled={!tippyContent}
+          placement="bottom"
+          content={tippyContent || <></>}
+          theme="light"
+          hideOnClick={false}
         >
-          <FormattedMessage {...messages.vote} />
-        </Button>
+          <Button
+            buttonStyle="primary"
+            iconSize="20px"
+            icon="vote-ballot"
+            onClick={onReaction}
+            id={voteButtonId}
+          >
+            <FormattedMessage {...messages.vote} />
+          </Button>
+        </Tippy>
       )}
     </>
   );
